pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Logush Ruslan",email = "phenixruslan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false

dependencies = [
    "openai (>=1.68.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
